requirements.txt
# Data Warehouse Local Free - Python Dependencies

# Orchestration
prefect>=2.14.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23

# Data processing
pyyaml>=6.0.1
python-dotenv>=1.0.0

# HTTP requests
requests>=2.31.0
httpx>=0.25.0

# Utilities
pydantic>=2.5.0

# Data Quality
great-expectations>=0.18.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.11.0
flake8>=6.1.0
isort>=5.12.0

